requirements.txt
vkbottle==4.3.12
openai==1.54.0
python-dotenv==1.0.0
aiosqlite==0.19.0
